requirements.txt
azure-ai-formrecognizer
azure-identity
fastapi
openai[datalib]>=1.3.7
tiktoken
tenacity
azure-ai-documentintelligence
azure-search-documents==11.6.0b1
azure-storage-blob
azure-storage-file-datalake
aiofiles
uvicorn
aiohttp
azure-monitor-opentelemetry
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-aiohttp-client
opentelemetry-instrumentation-openai
msal
azure-keyvault-secrets
cryptography
python-jose[cryptography]
types-python-jose
Pillow
types-Pillow
pypdf
PyMuPDF
beautifulsoup4
types-beautifulsoup4
# msgraph-sdk==1.1.0
ruff
black
pytest
pytest-asyncio
pytest-snapshot
coverage
playwright
pytest-cov
pytest-playwright
pytest-snapshot
PyMuPDF
pre-commit
locust
pip-tools
mypy
fpdf
azure-cosmos
PyPDF2